Switch the power menu to a Rofi menu script The power menu was initially handled by an i3 mode a basic `notify-send` notification.
As effective as it was, this method isn't user friendly and also "insecure" (in the sense that it sometimes wasn't so clear that you were in the power menu mode, which could lead to accidental suspend/reboot/shutdown).

The power menu is now handled via a simple bash script that spawns a rofi menu displaying the available options (lock, suspend, reboot...).
This is way more user friendly (search through options, mouse support, etc...) and more "secure" (as it is now clear that you opened the power menu, preventing accidental/undesired actions).

This new power menu also adds an option for the new [systemctl soft-reboot](https://lemm.ee/post/2464821) command.