Update Secure Boot setup (#229)
Add the signing of the /boot/EFI/systemd/systemd-bootx64.efi file (in addition of the already signed /usr/lib/systemd/boot/efi/systemd-bootx64.efi.signed file). This shouldn't be necessary usually but for some reason, it prevented a "Secure Boot violation" issue on some of my PCs.

Also add the signing of the vmlinuz kernel file (which is definitely not necessary since we boot the UKI instead, but it also doesn't hurt) and force the update of the bootloader after the sbctl setup, just in case.