Remove obsolete installation procedures (#269)
Gentoo's procedure needs to be revisited (some outdated parts or stuff I don't use anymore, but I have no plan to re-install Gentoo in a near future so let's just get rid of them for now.

As sad as it is, ALARM is in a pretty bad state and I wouldn't recommend to use it anymore. Let's get rid of it as well.